this fixes #312
